Move OutputWriters to output package
