+++ generator/build.gradle
@@ -4,6 +4,7 @@
 
 dependencies {
     compile project(':common')
+    compile project(':output')
 
     compile 'org.jetbrains.kotlin:kotlin-stdlib:1.3.21'
     compile 'com.google.inject:guice:4.1.0'
