+++ generator/src/main/java/com/scottlogic/deg/generator/guice/SingleDatasetOutputTargetProvider.java
@@ -3,8 +3,8 @@ package com.scottlogic.deg.generator.guice;
 import com.google.inject.Inject;
 import com.google.inject.Provider;
 import com.google.inject.name.Named;
-import com.scottlogic.deg.generator.outputs.formats.OutputWriterFactory;
-import com.scottlogic.deg.generator.outputs.formats.trace.TraceOutputWriterFactory;
+import com.scottlogic.deg.output.writer.OutputWriterFactory;
+import com.scottlogic.deg.output.writer.trace.TraceOutputWriterFactory;
 import com.scottlogic.deg.generator.outputs.targets.*;
 import com.scottlogic.deg.generator.utils.FileUtils;
