+++ generator/src/main/java/com/scottlogic/deg/generator/outputs/targets/SingleDatasetOutputTarget.java
@@ -1,7 +1,7 @@
 package com.scottlogic.deg.generator.outputs.targets;
 
 import com.scottlogic.deg.common.profile.ProfileFields;
-import com.scottlogic.deg.generator.outputs.formats.DataSetWriter;
+import com.scottlogic.deg.output.writer.DataSetWriter;
 
 import java.io.IOException;
