+++ generator/src/main/java/com/scottlogic/deg/generator/outputs/targets/SplittingOutputTarget.java
@@ -1,9 +1,8 @@
 package com.scottlogic.deg.generator.outputs.targets;
 
-import com.scottlogic.deg.common.profile.Profile;
 import com.scottlogic.deg.common.profile.ProfileFields;
-import com.scottlogic.deg.generator.outputs.formats.DataSetWriter;
-import com.scottlogic.deg.generator.outputs.formats.SplittingDataSetWriter;
+import com.scottlogic.deg.output.writer.DataSetWriter;
+import com.scottlogic.deg.output.writer.SplittingDataSetWriter;
 
 import java.io.IOException;
 import java.util.ArrayList;
