+++ generator/src/test/java/com/scottlogic/deg/generator/outputs/formats/csv/CsvOutputWriterFactoryTests.java
@@ -6,7 +6,8 @@ import com.scottlogic.deg.common.profile.Field;
 import com.scottlogic.deg.common.profile.ProfileFields;
 import com.scottlogic.deg.common.output.DataBagValue;
 import com.scottlogic.deg.common.output.GeneratedObject;
-import com.scottlogic.deg.generator.outputs.formats.DataSetWriter;
+import com.scottlogic.deg.output.writer.DataSetWriter;
+import com.scottlogic.deg.output.writer.csv.CsvOutputWriterFactory;
 import org.hamcrest.Matcher;
 import org.junit.Assert;
 import org.junit.jupiter.api.Test;
