+++ orchestrator/src/test/java/com/scottlogic/deg/orchestrator/cucumber/testframework/utils/InMemoryOutputTarget.java
@@ -3,7 +3,7 @@ package com.scottlogic.deg.orchestrator.cucumber.testframework.utils;
 import com.scottlogic.deg.common.profile.ProfileFields;
 import com.scottlogic.deg.common.output.DataBagValue;
 import com.scottlogic.deg.common.output.GeneratedObject;
-import com.scottlogic.deg.generator.outputs.formats.DataSetWriter;
+import com.scottlogic.deg.output.writer.DataSetWriter;
 import com.scottlogic.deg.generator.outputs.targets.SingleDatasetOutputTarget;
 
 import java.util.List;
