+++ output/src/main/java/com/scottlogic/deg/output/writer/SplittingDataSetWriter.java
@@ -1,6 +1,7 @@
-package com.scottlogic.deg.generator.outputs.formats;
+package com.scottlogic.deg.output.writer;
 
 import com.scottlogic.deg.common.output.GeneratedObject;
+import com.scottlogic.deg.output.writer.DataSetWriter;
 
 import java.io.Closeable;
 import java.io.IOException;
