+++ output/src/main/java/com/scottlogic/deg/output/writer/json/JsonOutputWriterFactory.java
@@ -1,8 +1,8 @@
-package com.scottlogic.deg.generator.outputs.formats.json;
+package com.scottlogic.deg.output.writer.json;
 
 import com.scottlogic.deg.common.profile.ProfileFields;
-import com.scottlogic.deg.generator.outputs.formats.DataSetWriter;
-import com.scottlogic.deg.generator.outputs.formats.OutputWriterFactory;
+import com.scottlogic.deg.output.writer.DataSetWriter;
+import com.scottlogic.deg.output.writer.OutputWriterFactory;
 
 import java.io.IOException;
 import java.io.OutputStream;
