+++ output/src/main/java/com/scottlogic/deg/output/writer/trace/SourceTracingDataSetWriter.java
@@ -1,4 +1,4 @@
-package com.scottlogic.deg.generator.outputs.formats.trace;
+package com.scottlogic.deg.output.writer.trace;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.ObjectWriter;
@@ -8,7 +8,7 @@ import com.scottlogic.deg.common.profile.constraints.atomic.AtomicConstraint;
 import com.scottlogic.deg.common.output.CellSource;
 import com.scottlogic.deg.common.output.GeneratedObject;
 import com.scottlogic.deg.common.output.RowSource;
-import com.scottlogic.deg.generator.outputs.formats.DataSetWriter;
+import com.scottlogic.deg.output.writer.DataSetWriter;
 
 import java.io.*;
 import java.util.Collection;
